server/requirements.txt
fastapi
uvicorn
mediapipe
opencv-python
tensorflow-cpu

gTTS
transformers
torch

sqlalchemy
psycopg2-binary
python-dotenv
alembic